Teal Module: box plot

This shiny module displays a box plot

tm_g_boxplot(label, dataname, param_var, param, param_choices = param,
  value_var = "AVAL", value_var_choices = c("AVAL", "CHG"),
  plot_height = c(600, 200, 2000), trt_group = "ARM", color_manual = NULL,
  shape_manual = NULL, trt_group_choices = NULL, facet_var = "ARM",
  facet_var_choices = NULL, xaxis_var = "AVISIT",
  xaxis_var_choices = NULL, loq_flag_var = NULL, hline = NULL,
  facet_ncol = NULL, rotate_xlab = FALSE, pre_output = NULL,
  post_output = NULL, armlabel = NULL, filter_vars = c("BASE2", "BASE"),
  filter_labs = c("Screening", "Baseline"), code_data_processing = NULL)

Arguments

label
menu item label of the module in the teal app

dataname
analysis data used in teal module, needs to be available in the list passed to the data argument of init. Note that the data are expected to be in vertical form with the PARAMCD variable filtering to one observation per patient per visit.

param_var
columm in dataname to select parameters

param
default parameter selected from param_var

param_choices
vector of choices to choose the parameter from

value_var
default columm in dataname to plot.

value_var_choices
choice of which columns in dataname to plot

plot_height
numeric vectors to define the plot height.

trt_group
treatment variable. The boxes and symbols are colored according to this parameter.

color_manual
vector of treatment colors. assigned values in app.R otherwise uses default colors.

shape_manual
vector of LOQ shapes. assigned values in app.R otherwise uses default shapes.

trt_group_choices
choices for trt_group. If not specified then trt_group cannot be changed.

facet_var
variable to facet the plots by.

facet_var_choices
choices for facet_var. if not specified then facet_var will be fixed and cannot be changed.

xaxis_var
variable to categorize the x-axis

xaxis_var_choices
choices for xaxis_var. if not specified then xaxis_var will be fixed and cannot be changed.

loq_flag_var
variable for the LOQ. Values are "Y" or "N"

hline
y-axis value to position a horizontal line. NULL = No line.

facet_ncol
numeric value indicating number of facets per row. NULL = Use the default for ggplot2::facet_wrap.

rotate_xlab
45 degree rotation of x-axis values.

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

armlabel
header for the treatment symbols in the legend. If not specified then the label attribute for trt_group will be used. If there is no label attribute for trt_group, then the name of the parameter ( in title case) will be used.

filter_vars
Variables to be used for filtering the data. The default is BASE2 and BASE

filter_labs
Labels for the radio buttons for the filter_vars. The defaults are "Screening" for BASE2 and "Baseline" for BASE.

code_data_processing
Not used

Value

an module object#'

Details

provide additional information as needed. link to specification file http://rstudio.com

Examples

# NOT RUN {
# Example using analysis dataset for example ASL or ADSL,
# ALB points to biomarker data stored in a typical LB structure. for example
# ALB or ADLB.

# Example using analysis dataset for example ASL or ADSL,
# ABM points to biomarker data stored in a custom file created to support
# goshawk. for example ADBIOM
library(dplyr)

# assign data frame note that this needs to be done once in the app.R file
# but should be available # here during testing

ASL <- ASL
ALB <- ALB

x <- teal::init(
  data =  list(ASL = ASL, ALB = ALB),
  modules = root_modules(
      tm_g_boxplot(
        label = "Box Plot",
        dataname = "ALB",
        param_var = "PARAMCD",
        param = "IGA",
        param_choices = c("CRP", "IGA", "IGG", "IGM"),
        value_var = "AVAL",
        value_var_choices = c("AVAL", "BASE", "CHG"),
        rotate_xlab = FALSE,
        xaxis_var = "AVISITCD",
        trt_group = "ARM"
      )
  )
)
shinyApp(x$ui, x$server)

# }