Density Distribution Plot

This teal module renders the UI and calls the functions that create a density distribution plot and an accompanying summary table.

tm_g_density_distribution_plot(label, dataname, param_var,
  param_choices = param, param, xaxis_var,
  xaxis_var_choices = xaxis_var, trt_group = "ARM",
  color_manual = NULL, color_comb = NULL, plot_height = c(500, 200,
  2000), font_size = c(12, 8, 20), line_size = c(1, 0.25, 3),
  hline = NULL, facet_ncol = 2, rotate_xlab = FALSE,
  pre_output = NULL, post_output = NULL, code_data_processing = NULL)

Arguments

label
menu item label of the module in the teal app.

dataname
analysis data passed to the data argument of teal init. E.g. ADaM structured laboratory data frame ALB.

param_var
name of variable containing biomarker codes e.g. PARAMCD.

param_choices
list of biomarkers of interest.

param
biomarker selected.

xaxis_var
name of variable containing biomarker results displayed on x-axis e.g. AVAL.

xaxis_var_choices
list of variables containing biomarker results choices.

trt_group
name of variable representing treatment group e.g. ARM.

color_manual
vector of colors applied to treatment values.

color_comb
name or hex value for combined treatment color.

plot_height
controls plot height.

font_size
font size control for title, x-axis label, y-axis label and legend.

line_size
plot line thickness.

hline
y-axis value to position a horizontal line.

facet_ncol
numeric value indicating number of facets per row.

rotate_xlab
45 degree rotation of x-axis values.

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

code_data_processing
TODO

Details

None

Examples

# NOT RUN {
# Example using ADaM structure analysis dataset.

library(dplyr)
library(ggplot)
library(random.cdisc.data)
library(stringr)

# original ARM value = dose value
arm_mapping <- list("A: Drug X" = "150mg QD", "B: Placebo" = "Placebo",
"C: Combination" = "Combination")
color_manual <-  c("150mg QD" = "#000000", "Placebo" = "#3498DB", "Combination" = "#E74C3C")
# assign LOQ flag symbols: circles for "N" and triangles for "Y", squares for "NA"
shape_manual <-  c("N"  = 1, "Y"  = 2, "NA" = 0)

ASL <- radsl(N = 20, seed = 1)
ALB <- radlb(ASL, visit_format = "WEEK", n_assessments = 7, seed = 2)
ALB <- ALB %>%
mutate(AVISITCD = case_when(
AVISIT == "SCREENING" ~ "SCR",
AVISIT == "BASELINE" ~ "BL", grepl("WEEK", AVISIT) ~ paste("W",trimws(substr(AVISIT, start=6,
stop=str_locate(AVISIT, "DAY")-1))),
TRUE ~ as.character(NA))) %>%
mutate(AVISITCDN = case_when(AVISITCD == "SCR" ~ -2,
AVISITCD == "BL" ~ 0, grepl("W", AVISITCD) ~ as.numeric(gsub("\\D+", "", AVISITCD)),
TRUE ~ as.numeric(NA))) %>%
# use ARMCD values to order treatment in visualization legend
mutate(TRTORD = ifelse(grepl("C", ARMCD), 1,
ifelse(grepl("B", ARMCD), 2,
ifelse(grepl("A", ARMCD), 3, NA)))) %>%
mutate(ARM = as.character(arm_mapping[match(ARM, names(arm_mapping))])) %>%
mutate(ARM = factor(ARM) %>% reorder(TRTORD))

param_choices = c("ALT", "CRP", "IGA")

x <- teal::init(
  data = list(ASL = ASL, ALB = ALB),
  modules = root_modules(
    tm_g_density_distribution_plot(
       label = "Density Distribution Plot",
       dataname = "ALB",
       param_var = "PARAMCD",
       param_choices = param_choices,
       param = param_choices[1],
       xaxis_var = "AVAL",
       xaxis_var_choices = c("AVAL", "BASE", "CHG", "PCHG", "AVALL2"),
       trt_group = "ARM",
       color_manual = color_manual,
       color_comb = "#39ff14",
       plot_height = c(500, 200, 2000),
       font_size = c(12, 8, 20),
       line_size = c(1, .25, 3)
   )
  )
)

shinyApp(x$ui, x$server)

# }