Function to create a correlation plot without visit facetting. Presents all visit data based on analysis day.

This teal module renders the UI and calls the function that creates a correlation plot.

tm_g_gh_correlationplot_av(
  label,
  dataname,
  param_var,
  xaxis_param = xaxis_param,
  xaxis_param_choices = xaxis_param,
  xaxis_var = xaxis_var,
  xaxis_var_choices = xaxis_var,
  yaxis_param = yaxis_param,
  yaxis_param_choices = yaxis_param,
  yaxis_var = yaxis_var,
  yaxis_var_choices = yaxis_var,
  trt_group = "ARM",
  color_manual = NULL,
  shape_manual = NULL,
  facet_ncol = 2,
  facet_var = "ACTARM",
  reg_line = FALSE,
  rotate_xlab = FALSE,
  hline = NULL,
  vline = NULL,
  plot_height = c(500, 200, 2000),
  font_size = c(12, 8, 20),
  dot_size = c(1, 1, 12),
  reg_text_size = c(3, 3, 10),
  pre_output = NULL,
  post_output = NULL
)

Arguments

label
menu item label of the module in the teal app.

dataname
analysis data passed to the data argument of teal init. E.g. ADaM structured laboratory data frame ADLB.

param_var
name of variable containing biomarker codes e.g. PARAMCD.

xaxis_param
biomarker selected.

xaxis_param_choices
list of biomarkers of interest.

xaxis_var
name of variable containing biomarker results displayed on x-axis e.g. BASE.

xaxis_var_choices
list of variables containing biomarker results choices.

yaxis_param
biomarker selected.

yaxis_param_choices
list of biomarkers of interest.

yaxis_var
name of variable containing biomarker results displayed on y-axis e.g. AVAL.

yaxis_var_choices
list of variables containing biomarker results choices.

trt_group
name of variable representing treatment group e.g. ARM.

color_manual
vector of colors applied to treatment values.

shape_manual
vector of symbols applied to LOQ values.

facet_ncol
numeric value indicating number of facets per row.

facet_var
variable to use for treatment facetting.

reg_line
include regression line and annotations for slope and coefficient in visualization. Use with facet TRUE.

rotate_xlab
45 degree rotation of x-axis values.

hline
y-axis value to position of horizontal line.

vline
x-axis value to position a vertical line.

plot_height
controls plot height.

font_size
font size control for title, x-axis label, y-axis label and legend.

dot_size
plot dot size.

reg_text_size
font size control for regression line annotations.

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example a title.

post_output
optional, shiny.tag with text placed after the output to put the output into context. For example the helpText elements are useful.

Details

This module displays a correlation plot.

Examples

if (FALSE) {

# Example using ADaM structure analysis dataset.

library(random.cdisc.data)

# original ARM value = dose value
arm_mapping <- list("A: Drug X" = "150mg QD",
                    "B: Placebo" = "Placebo",
                    "C: Combination" = "Combination")

ADSL <- radsl(N = 20, seed = 1)
ADLB <- radlb(ADSL, visit_format = "WEEK", n_assessments = 7L, seed = 2)
ADLB <- ADLB %>%
  mutate(AVISITCD = case_when(
    AVISIT == "SCREENING" ~ "SCR",
    AVISIT == "BASELINE" ~ "BL",
    grepl("WEEK", AVISIT) ~ paste("W", stringr::str_extract(AVISIT, "(?<=(WEEK ))[0-9]+")),
    TRUE ~ as.character(NA)),
    AVISITCDN = case_when(
      AVISITCD == "SCR" ~ -2,
      AVISITCD == "BL" ~ 0,
      grepl("W", AVISITCD) ~ as.numeric(gsub("[^0-9]*", "", AVISITCD)),
      TRUE ~ as.numeric(NA)),
    TRTORD = case_when(
      ARMCD == "ARM C" ~ 1,
      ARMCD == "ARM B" ~ 2,
      ARMCD == "ARM A" ~ 3),
    ARM = as.character(arm_mapping[match(ARM, names(arm_mapping))]),
    ARM = factor(ARM) %>% reorder(TRTORD),
    ADY = AVISITCDN)

x <- teal::init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADLB", ADLB),
    code =
      [1069 chars quoted with '''],
    check = FALSE
  ),
  modules = root_modules(
    tm_g_gh_correlationplot_av(
      label = "Correlation Plot",
      dataname = "ADLB",
      param_var = "PARAMCD",
      xaxis_param_choices = c("ALT", "CRP", "IGA"),
      xaxis_param = "ALT",
      xaxis_var = "BASE",
      xaxis_var_choices = c("AVAL", "BASE", "CHG", "PCHG", "BASE2", "AVALL2"),
      yaxis_param_choices = c("ALT", "CRP", "IGA"),
      yaxis_param = "CRP",
      yaxis_var = "AVAL",
      yaxis_var_choices = c("AVAL", "BASE", "CHG", "PCHG", "BASE2", "AVALL2"),
      trt_group = "ACTARM",
      color_manual = c("150mg QD" = "#000000",
                       "Placebo" = "#3498DB",
                       "Combination" = "#E74C3C"),
      shape_manual = c("N"  = 1, "Y"  = 2, "NA" = 0),
      plot_height = c(500, 200, 2000),
      facet_ncol = 2,
      facet_var = "ADY",
      reg_line = FALSE,
      font_size = c(12, 8, 20),
      dot_size = c(1, 1, 12),
      reg_text_size = c(3, 3, 10)
    )
  )
)

shinyApp(x$ui, x$server)

}