Box Plot

This teal module renders the UI and calls the functions that create a box plot and accompanying summary table.

tm_g_boxplot(label, dataname, param_var, param_choices = param, param,
  yaxis_var = "AVAL", yaxis_var_choices = c("AVAL", "CHG"),
  xaxis_var = "AVISITCD", xaxis_var_choices = NULL,
  facet_var = "ARM", facet_var_choices = NULL,
  filter_vars = c("BASE2", "BASE"), filter_labs = c("Screening",
  "Baseline"), trt_group = "ARM", color_manual = NULL,
  shape_manual = NULL, hline = NULL, facet_ncol = NULL,
  rotate_xlab = FALSE, pre_output = NULL, post_output = NULL,
  armlabel = NULL, plot_height = c(600, 200, 2000),
  code_data_processing = NULL)

Arguments

label
menu item label of the module in the teal app.

dataname
analysis data passed to the data argument of teal init. E.g. ADaM structured laboratory data frame ALB.

param_var
name of variable containing biomarker codes e.g. PARAMCD.

param_choices
list of biomarkers of interest.

param
biomarker selected.

yaxis_var
name of variable containing biomarker results displayed on y-axis e.g. AVAL.

yaxis_var_choices
list of variables containing biomarker results choices.

xaxis_var
variable to categorize the x-axis.

xaxis_var_choices
variable choices with which to categorize x-axis.

facet_var
variable to facet the plots by.

facet_var_choices
variable choices with which to facet the plots.

filter_vars
Variables to be used for filtering the data. The default is BASE2 and BASE

filter_labs
Labels for the radio buttons for the filter_vars. The defaults are "Screening" for BASE2 and "Baseline" for BASE.

trt_group
name of variable representing treatment group e.g. ARM.

color_manual
vector of colors applied to treatment values.

shape_manual
vector of symbols applied to LOQ values.

hline
y-axis value to position a horizontal line. NULL = No line.

facet_ncol
numeric value indicating number of facets per row.

rotate_xlab
45 degree rotation of x-axis values.

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

armlabel
label for the treatment symbols in the legend. If not specified then the label attribute for trt_group will be used. If there is no label attribute for trt_group, then the name of the parameter (in title case) will be used.

plot_height
numeric vectors to define the plot height.

code_data_processing
TODO

Value

an module object#'

Examples

# NOT RUN {
# Example using ADaM structure analysis dataset.
# ALB refers to biomarker data stored in expected laboratory structure.

library(dplyr)

ASL <- ADSL
ALB <- ADLB

x <- teal::init(
  data =  list(ASL = ASL, ALB = ALB),
  modules = root_modules(
      tm_g_boxplot(
        label = "Box Plot",
        dataname = "ALB",
        param_var = "PARAMCD",
        param = "CRP",
        param_choices = c("CRP", "IGA", "IGG", "IGM"),
        yaxis_var = "AVAL",
        yaxis_var_choices = c("AVAL", "BASE", "CHG"),
        rotate_xlab = FALSE,
        xaxis_var = "ARM",
        xaxis_var_choices = c("ARM", "AVISITCD", "STUDYID"),
        facet_var= "AVISITCD",
        facet_var_choices = c("ARM", "AVISITCD"),
        trt_group = "ARM"
      )
  )
)
shinyApp(x$ui, x$server)

# }