Spaghetti Plot

This teal module renders the UI and calls the function that creates a spaghetti plot.

tm_g_gh_spaghettiplot(label, dataname, param_var, param,
  param_var_label = "PARAM", idvar, xaxis_var, yaxis_var,
  xaxis_var_level = NULL, filter_var = yaxis_var, trt_group,
  trt_group_level = NULL, group_stats = "NONE", hline = NULL,
  man_color = NULL, color_comb = NULL, xtick = waiver(),
  xlabel = xtick, rotate_xlab = FALSE, facet_ncol = 2,
  plot_height = c(600, 200, 2000), font_size = c(12, 8, 20),
  pre_output = NULL, post_output = NULL)

Arguments

label
menu item label of the module in the teal app.

dataname
analysis data passed to the data argument of teal init. E.g. ADaM structured laboratory data frame ADLB.

param_var
name of variable containing biomarker codes e.g. PARAMCD.

param
biomarker selected.

param_var_label
single name of variable in analysis data that includes parameter labels.

idvar
name of unique subject id variable.

xaxis_var
single name of variable in analysis data that is used as x-axis in the plot for the respective goshawk function.

yaxis_var
single name of variable in analysis data that is used as summary variable in the respective gshawk function.

xaxis_var_level
vector that can be used to define the factor level of xaxis_var. Only use it when xaxis_var is character or factor.

filter_var
data constraint variable.

trt_group
name of variable representing treatment group e.g. ARM.

trt_group_level
vector that can be used to define factor level of trt_group.

group_stats
control group mean or median overlay.

hline
numeric value to add horizontal line to plot

man_color
string vector representing customized colors

color_comb
name or hex value for combined treatment color.

xtick
numeric vector to define the tick values of x-axis when x variable is numeric. Default value is waive().

xlabel
vector with same length of xtick to define the label of x-axis tick values. Default value is waive().

rotate_xlab
boolean value indicating whether to rotate x-axis labels

facet_ncol
numeric value indicating number of facets per row.

plot_height
numeric vectors to define the plot height.

font_size
control font size for title, x-axis, y-axis and legend font.

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

Value

shiny object

Examples

# Example using ADaM structure analysis dataset.

library(random.cdisc.data)

# original ARM value = dose value
arm_mapping <- list("A: Drug X" = "150mg QD",
                    "B: Placebo" = "Placebo",
                    "C: Combination" = "Combination")

ADSL <- radsl(cached = TRUE)
ADLB <- radlb(cached = TRUE)
ADLB <- ADLB %>%
  mutate(AVISITCD = case_when(
    AVISIT == "SCREENING" ~ "SCR",
    AVISIT == "BASELINE" ~ "BL",
    grepl("WEEK", AVISIT) ~ paste("W", stringr::str_extract(AVISIT, "(?<=(WEEK ))[0-9]+")),
    TRUE ~ as.character(NA)),
    AVISITCDN = case_when(
      AVISITCD == "SCR" ~ -2,
      AVISITCD == "BL" ~ 0,
      grepl("W", AVISITCD) ~ as.numeric(gsub("[^0-9]*", "", AVISITCD)),
      TRUE ~ as.numeric(NA)),
    AVISITCD = factor(AVISITCD) %>% reorder(AVISITCDN),
    TRTORD = case_when(
      ARMCD == "ARM C" ~ 1,
      ARMCD == "ARM B" ~ 2,
      ARMCD == "ARM A" ~ 3),
    ARM = as.character(arm_mapping[match(ARM, names(arm_mapping))]),
    ARM = factor(ARM) %>% reorder(TRTORD))

app <- teal::init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    cdisc_dataset("ADLB", ADLB),
    code = {[1064 chars quoted with ''']},
    check = FALSE
  ),
  modules = root_modules(
    tm_g_gh_spaghettiplot(
      label = "Spaghetti Plot",
      dataname = "ADLB",
      param_var = "PARAMCD",
      param = choices_selected(c("ALT", "CRP", "IGA"), "ALT"),
      idvar = "USUBJID",
      xaxis_var = choices_selected(c("Analysis Visit Code" = "AVISITCD"), "AVISITCD"),
      yaxis_var = choices_selected(c("AVAL","CHG", "PCHG"), "AVAL"),
      filter_var = choices_selected(c("None" = "NONE", "Screening" = "BASE2", "Baseline" = "BASE"),
       "NONE"),
      trt_group = "ARM",
      color_comb = "#39ff14",
      man_color = c('Combination' = "#000000",
                   'Placebo' = "#fce300",
                   '150mg QD' = "#5a2f5f")
    )
  )
)
if (FALSE) {
shinyApp(app$ui, app$server)
}